requirements.txt
streamlit==1.10.0
panda
